Day 9/services/linkedin_scraper.py
import requests
from bs4 import BeautifulSoup
import time
import random
from typing import Dict, List, Any
import logging
from config.settings import load_config

logger = logging.getLogger(__name__)

class LinkedInScraper:
    """LinkedIn scraper for alumni data (simulation for demo purposes)"""
    
    def __init__(self):
        """Initialize LinkedIn scraper"""
        self.config = load_config()
        self.session = requests.Session()
        
    def search_alumni(self, query: str) -> List[Dict[str, Any]]:
        """Search for alumni profiles (simulated data for demo)"""
        try:
            logger.info(f"Searching alumni with query: {query}")
            
            # Simulate network delay
            time.sleep(random.uniform(1, 3))
            
            # Generate simulated alumni data based on query
            alumni_data = self._generate_sample_alumni_data(query)
            
            logger.info(f"Found {len(alumni_data)} alumni profiles")
            return alumni_data
            
        except Exception as e:
            logger.error(f"Error searching alumni: {e}")
            return []
    
    def _generate_sample_alumni_data(self, query: str) -> List[Dict[str, Any]]:
        """Generate sample alumni data for demonstration"""
        
        companies = ["Google", "Microsoft", "Amazon", "Apple", "Meta", "Netflix", "Tesla", "Uber", "Airbnb", "Spotify"]
        roles = ["Software Engineer", "Senior Software Engineer", "Staff Engineer", "Engineering Manager", 
                "Data Scientist", "Product Manager", "DevOps Engineer", "ML Engineer", "Tech Lead"]
        skills = ["Python", "Java", "React", "Node.js", "AWS", "Docker", "Kubernetes", "Machine Learning", 
                 "TypeScript", "Go", "PostgreSQL", "MongoDB", "Redis", "GraphQL", "REST APIs"]
        
        sample_alumni = []
        
        # Parse query to influence generated data
        query_lower = query.lower()
        target_companies = [comp for comp in companies if comp.lower() in query_lower]
        if not target_companies:
            target_companies = companies[:5]  # Default selection
        
        # Generate 15-25 sample profiles
        num_profiles = random.randint(15, 25)
        
        for i in range(num_profiles):
            name_first = random.choice(["Alex", "Jordan", "Taylor", "Morgan", "Casey", "Riley", "Avery", "Quinn"])
            name_last = random.choice(["Smith", "Johnson", "Williams", "Brown", "Jones", "Garcia", "Miller", "Davis"])
            
            graduation_year = random.choice(range(2018, 2024))
            years_exp = 2025 - graduation_year - 1
            
            alumni = {
                "name": f"{name_first} {name_last}",
                "graduation_year": graduation_year,
                "current_company": random.choice(target_companies),
                "current_role": random.choice(roles),
                "years_of_experience": max(0, years_exp + random.randint(-1, 2)),
                "skills": random.sample(skills, random.randint(5, 10)),
                "industry": "Technology",
                "location": random.choice(["San Francisco, CA", "Seattle, WA", "New York, NY", "Austin, TX", "Boston, MA"]),
                "linkedin_url": f"https://linkedin.com/in/{name_first.lower()}-{name_last.lower()}-{random.randint(100, 999)}",
                "willing_to_refer": random.choice([True, True, True, False]),  # 75% willing
                "max_referrals_per_month": random.randint(2, 5),
                "referral_count_this_month": random.randint(0, 2),
                "email": f"{name_first.lower()}.{name_last.lower()}@{random.choice(['gmail.com', 'outlook.com', 'company.com'])}",
                "contact_preferences": {
                    "method": random.choice(["linkedin", "email"]),
                    "response_time": random.choice(["1-3 days", "2-5 days", "3-7 days"])
                },
                "company_history": [
                    {
                        "company": random.choice(companies),
                        "role": random.choice(roles[:4]),  # Earlier career roles
                        "duration": f"{random.randint(1, 3)} years"
                    }
                ]
            }
            
            sample_alumni.append(alumni)
        
        return sample_alumni
    
    def get_profile_details(self, profile_url: str) -> Dict[str, Any]:
        """Get detailed profile information (simulated)"""
        try:
            logger.info(f"Getting profile details for: {profile_url}")
            
            # Simulate network delay
            time.sleep(random.uniform(0.5, 2))
            
            # Extract name from URL for simulation
            name_part = profile_url.split('/')[-1] if profile_url else "unknown"
            
            profile_details = {
                "url": profile_url,
                "verified": random.choice([True, False]),
                "connections": random.randint(500, 2000),
                "recent_activity": [
                    "Posted about new project launch",
                    "Shared article about tech trends",
                    "Celebrated team achievement"
                ][:random.randint(1, 3)],
                "education": {
                    "university": "University Name",
                    "degree": "Bachelor of Science",
                    "major": random.choice(["Computer Science", "Software Engineering", "Data Science"]),
                    "graduation_year": random.randint(2018, 2023)
                },
                "certifications": random.sample([
                    "AWS Certified Solutions Architect",
                    "Google Cloud Professional",
                    "Certified Kubernetes Administrator",
                    "Scrum Master Certification"
                ], random.randint(0, 2))
            }
            
            return profile_details
            
        except Exception as e:
            logger.error(f"Error getting profile details: {e}")
            return {}
    
    def search_by_company(self, company_name: str) -> List[Dict[str, Any]]:
        """Search alumni by company"""
        try:
            query = f"alumni working at {company_name}"
            return self.search_alumni(query)
            
        except Exception as e:
            logger.error(f"Error searching by company {company_name}: {e}")
            return []
    
    def search_by_role(self, role: str) -> List[Dict[str, Any]]:
        """Search alumni by role"""
        try:
            query = f"alumni with role {role}"
            return self.search_alumni(query)
            
        except Exception as e:
            logger.error(f"Error searching by role {role}: {e}")
            return []
    
    def search_by_skills(self, skills: List[str]) -> List[Dict[str, Any]]:
        """Search alumni by skills"""
        try:
            skills_str = " ".join(skills)
            query = f"alumni with skills {skills_str}"
            return self.search_alumni(query)
            
        except Exception as e:
            logger.error(f"Error searching by skills: {e}")
            return []
    
    def get_company_employees(self, company_name: str, limit: int = 50) -> List[Dict[str, Any]]:
        """Get current employees at a company (simulated)"""
        try:
            logger.info(f"Getting employees for company: {company_name}")
            
            # Generate sample employee data
            employees = []
            num_employees = min(limit, random.randint(20, 40))
            
            for i in range(num_employees):
                employee = self._generate_sample_employee(company_name)
                employees.append(employee)
            
            return employees
            
        except Exception as e:
            logger.error(f"Error getting company employees: {e}")
            return []
    
    def _generate_sample_employee(self, company_name: str) -> Dict[str, Any]:
        """Generate a sample employee profile"""
        roles = ["Software Engineer", "Senior Software Engineer", "Principal Engineer", 
                "Engineering Manager", "Data Scientist", "Product Manager", "DevOps Engineer"]
        
        name_first = random.choice(["Sam", "Jamie", "Drew", "Blake", "Sage", "River", "Phoenix", "Rowan"])
        name_last = random.choice(["Wilson", "Anderson", "Taylor", "Moore", "Jackson", "Martin", "Lee", "White"])
        
        return {
            "name": f"{name_first} {name_last}",
            "company": company_name,
            "role": random.choice(roles),
            "tenure": f"{random.randint(1, 8)} years",
            "location": random.choice(["San Francisco, CA", "Seattle, WA", "New York, NY", "Remote"]),
            "profile_url": f"https://linkedin.com/in/{name_first.lower()}-{name_last.lower()}-{random.randint(100, 999)}",
            "is_alumni": random.choice([True, False]),
            "graduation_year": random.randint(2015, 2023) if random.choice([True, False]) else None
        }
    
    def close(self):
        """Close the scraper session"""
        try:
            self.session.close()
            logger.info("LinkedIn scraper session closed")
        except Exception as e:
            logger.error(f"Error closing scraper session: {e}")