Day 9/requirements.txt
streamlit>=1.28.0
pymongo>=4.6.0
google-generativeai>=0.3.0
crewai>=0.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
autogen-agentchat>=0.4.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
plotly>=5.17.0
aiohttp>=3.9.0
pydantic>=2.5.0
tenacity>=8.2.0
tiktoken>=0.5.0
autogen-agentchat>=0.4.0
protobuf>=5.0.0